refactor: PR 리뷰보고 리팩터링
