+++ src/main/java/com/ureca/juksoon/domain/user/dto/UserRoleReq.java
@@ -1,15 +1,11 @@
 package com.ureca.juksoon.domain.user.dto;
 
 import com.ureca.juksoon.domain.user.entity.UserRole;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Getter
-@Setter
 @NoArgsConstructor
-@AllArgsConstructor
 public class UserRoleReq {
     private UserRole userRole;
 }
