+++ src/main/java/com/ureca/juksoon/domain/user/dto/UserRoleRes.java
@@ -3,12 +3,10 @@ package com.ureca.juksoon.domain.user.dto;
 import com.ureca.juksoon.domain.user.entity.UserRole;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Getter
 @Setter
-@NoArgsConstructor
 @AllArgsConstructor
 public class UserRoleRes {
     private UserRole userRole;
