+++ src/main/java/com/ureca/juksoon/global/response/ResultCode.java
@@ -19,7 +19,7 @@ public enum ResultCode {
     UNSUPPORTED_REQUEST(HttpStatus.NOT_FOUND, 1007, "존재하지 않는 요청입니다."),
 
     // 사용자 2000번대
-
+    USER_NOT_FOUNT(HttpStatus.NOT_FOUND, 2000, "존재하지 않는 사용자입니다."),
     ;
 
     private final HttpStatus status;
