+++ src/main/java/com/ureca/juksoon/global/security/oauth/filter/CustomLogoutFilter.java
@@ -1,6 +1,6 @@
 package com.ureca.juksoon.global.security.oauth.filter;
 
-import com.ureca.juksoon.domain.refresh.service.RefreshTokenService;
+import com.ureca.juksoon.global.refresh.service.RefreshTokenService;
 import com.ureca.juksoon.global.response.CustomCookieType;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
